requirements.txt
groq
requests
feedparser
schedule
beautifulsoup4
python-dotenv
playwright
